Update Job instance and scheduled job id type from int to long
Standalone version will used a range of 1 000 000 prefix by an IP adresse leading to long id

+++ jqm-all/jqm-admin/src/main/java/com/enioka/api/admin/ScheduledJob.java
@@ -16,7 +16,7 @@ public class ScheduledJob implements Serializable
 {
     private static final long serialVersionUID = 7928212054684657247L;
 
-    private Integer id = null;
+    private Long id = null;
 
     private String cronExpression;
 
@@ -48,7 +48,7 @@ public class ScheduledJob implements Serializable
      *
      * @return the ID.
      */
-    public Integer getId()
+    public Long getId()
     {
         return this.id;
     }
@@ -58,7 +58,7 @@ public class ScheduledJob implements Serializable
      *
      * @param id
      */
-    public ScheduledJob setId(Integer id)
+    public ScheduledJob setId(Long id)
     {
         this.id = id;
         return this;
